@@ -1,47 +1,61 @@
 package ru.skypro.homework.service.impl;
 
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.mapper.UserResponseMapper;
+import ru.skypro.homework.model.User;
 import ru.skypro.homework.service.AuthService;
 
+@Slf4j
 @Service
+@RequiredArgsConstructor
 public class AuthServiceImpl implements AuthService {
 
-    private final UserDetailsManager manager;
+    private final UserDetailsService userDetailsService;
+    private final AuthenticationManager authenticationManager;
+    private final UserServiceImpl userServiceImpl;
+    private final UserResponseMapper userMapper;
     private final PasswordEncoder encoder;
-
-    public AuthServiceImpl(UserDetailsManager manager,
-                           PasswordEncoder passwordEncoder) {
-        this.manager = manager;
-        this.encoder = passwordEncoder;
-    }
 
     @Override
     public boolean login(String userName, String password) {
-        if (!manager.userExists(userName)) {
+        try {
+            Authentication authenticationToken = new UsernamePasswordAuthenticationToken(userName, password);
+            authenticationManager.authenticate(authenticationToken);
+            log.info("Авторизация пользователя: {} успешно пройдена", userName);
+            return true;
+        } catch (AuthenticationException e) {
+            log.error("Не удалось войти в систему для пользователя: {}", userName, e);
             return false;
         }
-        UserDetails userDetails = manager.loadUserByUsername(userName);
-        return encoder.matches(password, userDetails.getPassword());
     }
 
     @Override
     public boolean register(Register register) {
-        if (manager.userExists(register.getUsername())) {
+        try {
+            userDetailsService.loadUserByUsername(register.getUsername());
+            log.error(
+                    "Регистрация не удалась: пользователь с именем пользователя {} уже существует.",
+                    register.getUsername()
+            );
             return false;
+        } catch (UsernameNotFoundException e) {
+            User user = userMapper.fromRegister(register);
+            user.setPassword(encoder.encode(register.getPassword()));
+            userServiceImpl.save(user);
+            log.info(
+                    "Пользователь успешно зарегистрирован с именем пользователя: {}",
+                    register.getUsername());
+            return true;
         }
-        manager.createUser(
-                User.builder()
-                        .passwordEncoder(this.encoder::encode)
-                        .password(register.getPassword())
-                        .username(register.getUsername())
-                        .roles(register.getRole().name())
-                        .build());
-        return true;
     }
-
 }
